File breakout, rename to conform to e.g. xdr.Int128Parts
