+++ src/index.js
@@ -48,5 +48,6 @@ export {
 } from './util/decode_encode_muxed_account';
 
 export { ScInt, XdrInt } from './numbers/scint';
+export { Uint256, Int256, Uint128, Int128 } from './numbers/xdrint';
 
 export default module.exports;
